Rename mqtt mixins module to `entity.py`
