+++ homeassistant/components/mqtt/device_trigger.py
@@ -36,7 +36,7 @@ from .const import (
     DOMAIN,
 )
 from .discovery import MQTTDiscoveryPayload, clear_discovery_hash
-from .mixins import MqttDiscoveryDeviceUpdateMixin, send_discovery_done, update_device
+from .entity import MqttDiscoveryDeviceUpdateMixin, send_discovery_done, update_device
 from .models import DATA_MQTT
 from .schemas import MQTT_ENTITY_DEVICE_INFO_SCHEMA
