+++ homeassistant/components/mqtt/lock.py
@@ -34,7 +34,7 @@ from .const import (
     CONF_STATE_OPENING,
     CONF_STATE_TOPIC,
 )
-from .mixins import MqttEntity, async_setup_entity_entry_helper
+from .entity import MqttEntity, async_setup_entity_entry_helper
 from .models import (
     MqttCommandTemplate,
     MqttValueTemplate,
