+++ homeassistant/components/mqtt/siren.py
@@ -46,7 +46,7 @@ from .const import (
     PAYLOAD_EMPTY_JSON,
     PAYLOAD_NONE,
 )
-from .mixins import MqttEntity, async_setup_entity_entry_helper
+from .entity import MqttEntity, async_setup_entity_entry_helper
 from .models import (
     MqttCommandTemplate,
     MqttValueTemplate,
